+++ cmd/app/main.go
@@ -5,20 +5,18 @@ import (
 	"os"
 	"strings"
 
+	"github.com/dadez/file-upload/pkg/auth"
+	"github.com/dadez/file-upload/pkg/common"
+	"github.com/dadez/file-upload/pkg/healthz"
+	"github.com/dadez/file-upload/pkg/ui"
+	"github.com/dadez/file-upload/pkg/upload"
+
 	"github.com/gorilla/mux"
-	"github.com/sirupsen/logrus"
 	log "github.com/sirupsen/logrus"
 )
 
-const (
-	maxUploadSize = 1024 * 1024 // 1MB
-)
-
 var (
-	staticFilesPath      string
-	authFilesPath        string
-	uploadsDirectoryPath string
-	logLevel             string
+	logLevel string
 )
 
 func init() {
@@ -26,7 +24,7 @@ func init() {
 
 	//log.SetReportCaller(true)
 
-	formatter := &logrus.TextFormatter{
+	formatter := &log.TextFormatter{
 		FullTimestamp: true,
 	}
 
@@ -37,7 +35,7 @@ func init() {
 	log.SetOutput(os.Stdout)
 
 	// configure log level, default to info
-	logLevel = GetEnv("LOG_LEVEL", "info")
+	logLevel = common.GetEnv("LOG_LEVEL", "info")
 	lvl := strings.ToLower(logLevel)
 
 	switch lvl {
@@ -56,35 +54,17 @@ func init() {
 	case "panic":
 		log.SetLevel(log.PanicLevel)
 	}
-
-	// check for environment variables
-	// check for static files path
-	staticFilesPath = GetEnv("STATIC_FILES_PATH", ".") + "/"
-	// check for static files path
-	authFilesPath = GetEnv("AUTH_FILES_PATH", ".") + "/"
-	// check for uploads directory path
-	uploadsDirectoryPath = GetEnv("UPLOADS_DIRECTORY_PATH", "uploads")
-}
-
-// GetEnv checks for existing environment variables
-func GetEnv(key, fallback string) string {
-	value, exists := os.LookupEnv(key)
-	if !exists {
-		value = fallback
-	}
-	return value
 }
 
 // use gorilla mux for serve http
 func main() {
 	log.Info("fileupload server ready")
-	log.Debug("auth files under ", authFilesPath)
 	// run webserver
 	r := mux.NewRouter()
 
-	r.HandleFunc("/", uiHandler).Methods("GET")
-	r.HandleFunc("/healthz", healthzHandler).Methods("GET")
-	r.HandleFunc("/upload", BasicAuth(uploadHandler)).Methods("POST")
+	r.HandleFunc("/", ui.UiHandler).Methods("GET")
+	r.HandleFunc("/healthz", healthz.HealthzHandler).Methods("GET")
+	r.HandleFunc("/upload", auth.BasicAuth(upload.UploadHandler)).Methods("POST")
 
 	if err := http.ListenAndServe(":4500", r); err != nil {
 		log.Fatal(err)
